chore: reorganize grpc related code
