+++ src/entities/trader/web-taker.ts
@@ -1,4 +1,4 @@
-import type { ParsedQuoteResponse } from '../../utils';
+import type { ParsedQuoteResponse } from '../../grpc';
 import type { QuoteRequest } from '../../lib/codegen/rfq_pb';
 import { Trader } from './base-trader';
 import type { TraderConstructorArgs } from './base-trader';
