+++ src/grpc/hi-lo-bit-segmentation/hi-lo-to-big-int.ts
