+++ src/grpc/hi-lo-bit-segmentation/index.ts
