+++ src/grpc/hi-lo-bit-segmentation/parse-quote-response.ts
