+++ src/utils/index.ts
@@ -1,17 +1,2 @@
-export {
-  toH40,
-  toH96,
-  toH128,
-  toH160,
-  toH256,
-  fromH40,
-  fromH96,
-  fromH128,
-  fromH160,
-  fromH160ToAddress,
-  fromH256,
-  parseQuoteResponse,
-  type ParsedQuoteResponse,
-} from './hi-lo-bit-segmentation';
 export { createSIWEMessage } from './siwe';
 export { get24HrTimestamps, get8AMUTCDate } from './timestamps';
